+++ src/App.jsx
@@ -1,20 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
+import Todo from './Todo';
 
 export default function App() {
-  const [userInput, setUserInput] = useState({
-    text: '',
-  });
-
-  function handleChange(e) {
-    setUserInput({ text: e.target.value });
-  }
   return (
-    <p>
-      <h1>To-do</h1>
-      <input type="text" onChange={handleChange} />
-      <button type="button">
-        추가
-      </button>
-    </p>
+    <Todo />
   );
 }
